vessel/worker/processor: Retry reading from a repository
Attempt to re-read a file from a repository when the
previous attempt failed. The total number of attempts
is limited to 4.

Most common reason that the initial read fails is a
cache miss. Execute request can arrive before natsfs
is synchronized. Re-reading the file after some delay
(400 ms) should solve this issue.